+++ example/src/l18n/intl_config.ts
@@ -1,6 +1,6 @@
 import KTextConstants from "@/shared/constants/variables/text_constants";
-import { getCloudflareContext } from "@opennextjs/cloudflare";
 import { setIntlConfig } from "cloudflare-next-intl";
+import { getCloudflareContext } from "@opennextjs/cloudflare";
 
 declare global {
     type Language = "uk" | "en";
@@ -9,10 +9,22 @@ declare global {
 const intlConfig = setIntlConfig({
     locales: KTextConstants.locales,
     defaultLocale: KTextConstants.defaultLocale,
+    firebaseAuth: {
+        apiKey: process.env.NEXT_PUBLIC_FIREBASE_API_KEY!,
+        authDomain: process.env.NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN!,
+        projectId: process.env.NEXT_PUBLIC_FIREBASE_PROJECT_ID!,
+        storageBucket: process.env.NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET,
+        messagingSenderId: process.env.NEXT_PUBLIC_FIREBASE_MESSAGING_SENDER_ID,
+        appId: process.env.NEXT_PUBLIC_FIREBASE_APP_ID!,
+        measurementId: process.env.NEXT_PUBLIC_FIREBASE_MEASUREMENT_ID,
+        redirectAuthPath: '/login',
+        homePath: '/',
+        isAuthPath: (path: string) => path === '/login',
+    },
     cookieConsent: {
-        privacyPolicyDate: "2026-01-01",
+        privacyPolicyDate: "2026-01-05",
         getCloudflareContext: getCloudflareContext,
-        secrets: {}
+        analytics: {}
     },
 });
 export default intlConfig;